refactor(domain): make domain models pure and move I/O orchestration to Service
