+++ src/domain/models/status_list.rs
@@ -74,195 +74,6 @@ pub struct StatusListSnapshot {
     pub exp: i64,
 }
 
-use crate::domain::ports::{StatusListCache, StatusListHistoryRepo, StatusListRepo};
-
-fn current_unix_timestamp() -> i64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs() as i64
-}
-
-pub fn next_updated_at(previous: i64, now: i64) -> i64 {
-    now.max(previous + 1)
-}
-
-fn build_snapshot(record: &StatusListRecord, token_exp_secs: u64) -> StatusListSnapshot {
-    let iat = record.updated_at;
-    StatusListSnapshot {
-        snapshot_id: uuid::Uuid::new_v4().to_string(),
-        list_id: record.list_id.clone(),
-        issuer: record.issuer.clone(),
-        status_list: record.status_list.clone(),
-        sub: record.sub.clone(),
-        iat,
-        exp: iat + token_exp_secs as i64,
-    }
-}
-
-async fn invalidate_after_commit<C: StatusListCache + ?Sized>(cache: &C, list_id: &str) {
-    if let Err(_error) = cache.invalidate(list_id).await {
-        tracing::warn!(
-            list_id = %list_id,
-            error = ?_error,
-            "status list write committed, but cache invalidation failed; \
-             reads may be stale until the cache entry expires"
-        );
-    }
-}
-
-impl StatusListRecord {
-    /// Create and publish a new status list record, enforcing uniqueness and size invariants.
-    #[allow(clippy::too_many_arguments)]
-    pub async fn publish(
-        repo: &dyn StatusListRepo,
-        history_repo: Option<&dyn StatusListHistoryRepo>,
-        list_id: String,
-        issuer: Issuer,
-        sub: String,
-        statuses: Vec<StatusEntry>,
-        token_exp_secs: u64,
-        max_serialized_list_size: usize,
-    ) -> Result<Self, StatusListError> {
-        let record = Self {
-            list_id,
-            issuer,
-            status_list: StatusList::create(statuses)?,
-            sub,
-            updated_at: current_unix_timestamp(),
-        };
-
-        if record.status_list.lst.len() > max_serialized_list_size {
-            return Err(StatusListError::TooLarge);
-        }
-        if repo.find(&record.list_id).await?.is_some() {
-            return Err(StatusListError::AlreadyExists);
-        }
-
-        match history_repo {
-            Some(_) => {
-                let snapshot = build_snapshot(&record, token_exp_secs);
-                repo.insert_with_snapshot(record.clone(), snapshot).await?;
-            }
-            None => {
-                repo.insert(record.clone()).await?;
-            }
-        }
-
-        Ok(record)
-    }
-
-    /// Mutate statuses in an existing status list record with optimistic concurrency checks and cache invalidation.
-    #[allow(clippy::too_many_arguments)]
-    pub async fn update_statuses(
-        repo: &dyn StatusListRepo,
-        cache: &dyn StatusListCache,
-        history_repo: Option<&dyn StatusListHistoryRepo>,
-        issuer: &Issuer,
-        list_id: &str,
-        statuses: Vec<StatusEntry>,
-        token_exp_secs: u64,
-        max_status_index: i32,
-        max_statuses_per_request: usize,
-        max_serialized_list_size: usize,
-    ) -> Result<Self, StatusListError> {
-        if statuses.len() > max_statuses_per_request {
-            return Err(StatusListError::TooManyStatuses {
-                count: statuses.len(),
-                max: max_statuses_per_request,
-            });
-        }
-        for entry in &statuses {
-            if entry.index > max_status_index {
-                return Err(StatusListError::IndexTooLarge {
-                    index: entry.index,
-                    max: max_status_index,
-                });
-            }
-        }
-
-        let mut existing = repo
-            .find(list_id)
-            .await?
-            .ok_or(StatusListError::NotFound)?
-            .as_ref()
-            .clone();
-
-        if &existing.issuer != issuer {
-            return Err(StatusListError::IssuerMismatch);
-        }
-
-        existing.status_list = existing.status_list.update(statuses)?;
-        if existing.status_list.lst.len() > max_serialized_list_size {
-            return Err(StatusListError::TooLarge);
-        }
-
-        let previous_updated_at = existing.updated_at;
-        existing.updated_at = next_updated_at(previous_updated_at, current_unix_timestamp());
-
-        let landed = match history_repo {
-            Some(_) => {
-                let snapshot = build_snapshot(&existing, token_exp_secs);
-                repo.update_with_snapshot(existing.clone(), previous_updated_at, snapshot)
-                    .await?
-            }
-            None => repo.update(existing.clone(), previous_updated_at).await?,
-        };
-
-        if !landed {
-            return Err(StatusListError::Conflict);
-        }
-
-        invalidate_after_commit(cache, &existing.list_id).await;
-        Ok(existing)
-    }
-
-    /// Retrieve a status list record from cache or fallback to persistent storage.
-    pub async fn get(
-        repo: &dyn StatusListRepo,
-        cache: &dyn StatusListCache,
-        list_id: &str,
-    ) -> Result<Self, StatusListError> {
-        if let Some(record) = cache.get(list_id).await? {
-            return Ok(record.as_ref().clone());
-        }
-        let record = repo.find(list_id).await?.ok_or(StatusListError::NotFound)?;
-        cache.put(record.as_ref().clone()).await?;
-        Ok(record.as_ref().clone())
-    }
-
-    /// List all published status list URIs.
-    pub async fn list_uris(repo: &dyn StatusListRepo) -> Result<Vec<String>, StatusListError> {
-        repo.list_uris().await
-    }
-}
-
-impl StatusListSnapshot {
-    /// Retrieve a historical status list snapshot active at a specific Unix timestamp.
-    pub async fn get_valid_at(
-        history_repo: Option<&dyn StatusListHistoryRepo>,
-        list_id: &str,
-        time: i64,
-    ) -> Result<Self, StatusListError> {
-        history_repo
-            .ok_or(StatusListError::NotFound)?
-            .find_valid_at(list_id, time)
-            .await?
-            .ok_or(StatusListError::HistoricalNotFound)
-    }
-
-    /// Purge historical snapshots older than `cutoff` timestamp.
-    pub async fn cleanup(
-        history_repo: Option<&dyn StatusListHistoryRepo>,
-        cutoff: i64,
-    ) -> Result<u64, StatusListError> {
-        let Some(history) = history_repo else {
-            return Ok(0);
-        };
-        history.delete_older_than(cutoff).await
-    }
-}
-
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct StatusList {
     pub bits: u8,
